Safari-611.1.10/LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-en-GB.html

<!DOCTYPE html><!-- webkit-test-runner [ language=en-GB ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "en-gb");
</script>
</body>
</html>
